refactor(atls-ui-generators): flat namings
