+++ generators/icons/src/generator.ts
