+++ generators/icons/src/interfaces.ts
