+++ generators/icons/src/replacement-utils/create-replacement-icon-colors/constants.ts
