+++ generators/icons/src/replacement-utils/create-replacements/create-replacements.cts
