+++ generators/icons/src/replacement-utils/create-replacements/index.cts
@@ -1,1 +1,1 @@
-export * from './create-replacements.util.cjs'
+export * from './create-replacements.js'
