+++ generators/icons/src/replacement-utils/create-replacements/index.ts
@@ -1,1 +1,1 @@
-export * from './create-replacements.util.js'
+export * from './create-replacements.js'
