+++ generators/icons/src/replacement-utils/get-color-replacement/get-color-replacement.cts
