+++ generators/icons/src/replacement-utils/get-color-replacement/index.ts
@@ -1,1 +1,1 @@
-export * from './get-color-replacement.util.js'
+export * from './get-color-replacement.js'
