+++ generators/icons/src/replacement-utils/get-color-replacement/interfaces.ts
