+++ generators/icons/src/replacement-utils/write-replacements-file/write-replacements-file.ts
@@ -1,6 +1,6 @@
 /* eslint-disable n/no-sync */
 
-import type { Replacements } from '../../icons.interfaces.js'
+import type { Replacements } from '../../interfaces.js'
 
 import { writeFileSync }     from 'node:fs'
