+++ generators/icons/src/svgr-utils/get-svg-colors/config/constants.ts
