+++ generators/icons/src/svgr-utils/get-svg-colors/helpers/normalize-attribute-name.ts
@@ -1,4 +1,4 @@
-import { NON_WORD_PATTERN } from './regex.js'
+import { NON_WORD_PATTERN } from '../config/regex.js'
 
 export const normalizeAttributeName = (name: string): string =>
   name.replace(NON_WORD_PATTERN, '').toLowerCase()
