+++ generators/icons/src/svgr-utils/get-svg-colors/index.ts
@@ -1,2 +1,2 @@
-export * from './get-svg-colors.util.js'
-export type * from './get-svg-colors.interfaces.js'
+export * from './get-svg-colors.js'
+export type * from './interfaces.js'
