+++ generators/icons/src/svgr-utils/get-svg-colors/interfaces.ts
