+++ generators/icons/src/svgr-utils/index.ts
@@ -1,4 +1,4 @@
-export * from './build-svgr.util.js'
+export * from './build-svgr.js'
 export * from './compile-icons/index.js'
-export * from './discover-icons.util.js'
+export * from './discover-icons.js'
 export * from './get-svg-colors/index.js'
